Changed static front page gallery name and adjusted CSS to allow for line break.

+++ routes/admin.js
@@ -230,7 +230,7 @@ function updatePhoto(imageId, imageData) {
                 }); 
         
         });
-    } else {
+    } else if(imageData.deleteImage == 'true') {
         Photo.findByIdAndRemove(imageId, (err, doc) => {
             if(err) throw err;
             fs.unlink(`./public${doc.url.full}`, err => {
